Add a Performance Console - Move almost all of the implementation of 'Console.java' to 'GenericConsole.java', make 'GenericConsole' the new parent class for 'Console.java'. Now 'PerformanceConsole.java' which also inherits from 'GenericConsole' will have similar functionality as 'Console.java'
